backend/ml_service/requirements.txt
# ML Service (FastAPI) Requirements
fastapi==0.131.0
uvicorn[standard]==0.41.0
python-multipart==0.0.20
Pillow==10.4.0
numpy==2.4.2
opencv-python-headless==4.13.0.86
scikit-image==0.26.0
scipy==1.17.1
httpx==0.27.0
python-dotenv==1.2.1

# Production
gunicorn==21.2.0
supabase==2.9.0
postgrest==0.17.0
